--- components/webui/client/src/pages/SearchPage/SearchControls/Dataset/DatasetLabel/index.module.css
@@ -1,9 +1,0 @@
-.datasetLabel {
-    min-width: 80px;
-    height: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    border-top-right-radius: 0px;
-    border-bottom-right-radius: 0px;
-}
