+++ components/webui/client/src/pages/SearchPage/SearchControls/Presto/GuidedControls/OrderBy.tsx
@@ -1,0 +1,12 @@
+import InputLabel from "../../../../../components/InputLabel";
+import SqlInput from "../../../../../components/SqlInput";
+import guidedGrid from "./index.module.css";
+
+const OrderBy = () => (
+    <div className={guidedGrid["order"]}>
+        <InputLabel> ORDER BY </InputLabel>
+        <SqlInput disabled={false} />
+    </div>
+);
+
+export default OrderBy;
