+++ components/webui/client/src/pages/SearchPage/SearchControls/Presto/GuidedControls/Where.tsx
@@ -1,0 +1,12 @@
+import InputLabel from "../../../../../components/InputLabel";
+import SqlInput from "../../../../../components/SqlInput";
+import guidedGridStyles from "./index.module.css";
+
+const Where = () => (
+    <div className={guidedGridStyles["where"]}>
+        <InputLabel> WHERE </InputLabel>
+        <SqlInput disabled={false} />
+    </div>
+);
+
+export default Where;
